+++ CHANGELOG.md
@@ -9,6 +9,8 @@
 
 ### Fixed
 - Fixed the compression when performing conversions on PNG, that drastically reduces the file size, now it's set to `1`
+- Fixed undefined variable in `customer-new-order.blade.php`
+- Fixed automatic route-model binding for non-Bazar routes
 
 ### Removed
 - Removed the `Bazar\Repositories\Conversion` class